refactor: group cloud and analytics helpers
