+++ README.md
@@ -13,3 +13,8 @@ Import subpackages, for example:
 ```go
 import "github.com/patdeg/common/auth"
 ```
+You can also import specialized helpers:
+```go
+import "github.com/patdeg/common/gcp"
+import "github.com/patdeg/common/ga"
+```
