+++ gcp/appengine.go
@@ -1,4 +1,4 @@
-package common
+package gcp
 
 import (
 	"strings"
