+++ track/bigquery_helpers.go
@@ -2,6 +2,7 @@ package track
 
 import (
 	"github.com/patdeg/common"
+	"github.com/patdeg/common/gcp"
 
 	"golang.org/x/net/context"
 	bigquery "google.golang.org/api/bigquery/v2"
@@ -11,7 +12,7 @@ import (
 // insertWithTableCreation streams data to BigQuery and creates the table if it doesn't exist.
 func insertWithTableCreation(c context.Context, projectID, datasetID, tableID string, req *bigquery.TableDataInsertAllRequest, createTable func(context.Context, string) error) error {
 	common.Debug("insertWithTableCreation dataset=%s table=%s", datasetID, tableID)
-	err := common.StreamDataInBigquery(c, projectID, datasetID, tableID, req)
+	err := gcp.StreamDataInBigquery(c, projectID, datasetID, tableID, req)
 	if err != nil {
 		common.Error("Error while streaming data to BigQuery: %v", err)
 		if gerr, ok := err.(*googleapi.Error); ok && gerr.Code == 404 {
@@ -21,7 +22,7 @@ func insertWithTableCreation(c context.Context, projectID, datasetID, tableID st
 				common.Error("Error creating table %s: %v", tableID, err2)
 				return err2
 			}
-			if err3 := common.StreamDataInBigquery(c, projectID, datasetID, tableID, req); err3 != nil {
+			if err3 := gcp.StreamDataInBigquery(c, projectID, datasetID, tableID, req); err3 != nil {
 				common.Error("Error streaming to BigQuery after creating table: %v", err3)
 				return err3
 			}
